+++ tauri-plugin-screen-secure/build.rs
@@ -1,0 +1,11 @@
+// SPDX-FileCopyrightText: 2026 Zexin Yuan <gpm@yzx9.xyz>
+//
+// SPDX-License-Identifier: Apache-2.0
+
+const COMMANDS: &[&str] = &["set_secure"];
+
+fn main() {
+    tauri_plugin::Builder::new(COMMANDS)
+        .android_path("android")
+        .build();
+}